add ability to handle parsing multiple formats
added optional config attribute called altFormats; it is an array which is empty by default. 
On initialization, the value of formats is appended to the altFormats array.
The original option named format will continue to be the target output format, but the full array of altFormats will be used for validation and parsing.