Add customizable position option
This few lines of code allow you to have complete control on the
position of the calendar, relative to the field. However, automatic
adjustment will prevent the datepicker from being displayed outside the
viewport.
So we can consider this option as a preferred default position.

This commit include a pretty solid page of examples.