Fix issue #93 wrong month sometimes shown
The wrong month is sometimes shown when minDate  year > default date's

From the issue's description:
When you set minDate for a date after the currently selected date, the calendar, when shown, should display the month where minDate is. This works if both dates are in the same year, but if the new minDate's year is greater than the selected one, the month shown will be a mix of the selected date's month with the minDate's year.

There is a separate issue in that `setMinDate` and `setMaxDate` do not
update the min/maxMonth and min/maxYear internal options.