Fix Date.parse bug when using Datejs
Datejs replaces the native Date.parse method with one that understands
basic natual language. Unfortunately, it returns a Date object or null
on error unlike the native Date.parse method that returns milliseconds
since the UNIX epoch 1/1/1970 or NaN on error that Pikaday relied on.

- I have updated the parseDate method to check the result of
Date.parse instead of using it immediately in a new Date constructor
- A Datejs example has been added and the README updated accordingly
- Sugar.js example also updated to support format option and remove
mention of customising single Pikaday instance due to bug with Pikaday
constructor parsing default value string from field with normal
parseDate before it could be updated by user with this method. This
would require a new config option for custom formatter and parser for
Pikaday instance.
- README also updated to fix Knockout link and Javascript renamed to
JavaScript for consistency