Fix for #224, #230 in IE 7-11
In IE, a click in the pika div will make the input field lose focus. Previously, _onInputBlur did not test for this, so a click in the pika div would result in it closing.

This commit adds a check to see if document.activeElement is within the pika div, and if so does not allow pika to hide. IE sets this variable to the newly-focused element before the blur event fires;

IE sets document.activeElement to the newly-focused element before the blur event fires; this commit adds a check to see if the clicked element is within the pika div, and if so ignores the blur event. Other browsers correctly set the focused element to body before blur, so will not be affected by this change.

Tested in IE 7 (Win XP), 8-10 (Win 7), 11 (Win 8), latest firefox, chrome, opera