Allow passing array of formats for format option
If moment is installed it can accept an array of formats that allows for
parsing different user inputs into a valid date. This commit modifies
the `toString()` method so that if `format` is an array, it will use the
first item. This allows us to parse different formats from user input,
but always display a consistent format.