change today
change which day "today" is.

Use case : A service the website provides needs to use local dates+times and is based in Honk-Kong. You have a lot of users that will consume this service but are in West coast America and as such the days are potentially different at the time of access.