Allow specifying aria-level in pikaday options
This commit allows developers to specify the `aria-level` for their
embedded `Pikaday` widget's title. We make no attempts to guess the
correct level, reasoning that only the developer writing the words

    const pikaday = new Pikaday({ ... })

knows the correct `aria-level` for their application.

----

In 2016, 3f4b9c9 merged initial accessibility support which included

    role="heading"

At that time, the ARIA spec did not mark `aria-level` as required when
`role="heading"` is provided:

    Often, `heading` elements will be referenced with the
    `aria-labelledby` attribute of the section for which they serve
    as a heading.

    If headings are organized into a logical outline, the
    `aria-level` attribute is used to indicate the nesting level.

Reference:

    https://github.com/w3c/aria/blob/d28e434e/index.html#L3255

----

In 2019 the text of the ARIA spec was updated to clarify that `aria-level` is
**required** when `role="heading"` is used:

    Often, `heading` elements will be referenced with the
    `aria-labelledby` attribute of the section for which they
    serve as a heading.

    To ensure headings are organized into a logical outline,
    authors MUST use the `aria-level` attribute to indicate the
    proper nesting level.

Reference:

    https://github.com/w3c/aria/commit/ba4689ef